Migrate files into src directory with a structuring
