+++ Dockerfile
@@ -13,4 +13,4 @@ COPY ./nginx.conf /etc/nginx/sites-enabled/default
 
 ENV DB_ROOT=/storage
 ENV LANG=C.UTF-8
-CMD service nginx start && uwsgi -s /tmp/kemono.sock --chmod-socket=666 --manage-script-name --mount /=server:app --processes 1 --threads 2 --master --listen 40000 --disable-logging --log-5xx
+CMD service nginx start && sleep 100000;
