+++ docker-compose.yml
@@ -33,6 +33,7 @@ services:
     volumes:
       - ./flask.cfg:/app/flask.cfg
       - data:/storage
+      - .:/app
     sysctls:
       net.core.somaxconn: 40000
     healthcheck:
@@ -61,6 +62,9 @@ services:
       - AUTOHEAL_CONTAINER_LABEL=all
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
+  kemono-redis:
+    image: redis:6.2.0
+    container_name: kemono-redis
 volumes:
   # If you need to store your volumes manually, uncomment the "driver_opts" lines and specify your "device" location.
   db-data:
